Move language swicher component to components subfolder
