+++ resources/views/components/ui/mobile-sidebar.blade.php
@@ -27,7 +27,7 @@
                     <img class="h-8 w-auto" src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=500"
                         alt="Your Company">
                 </div>
-                @livewire('language-switcher')
+                @livewire('components.language-switcher')
                 <nav class="flex flex-1 flex-col">
                     <ul role="list" class="flex flex-1 flex-col gap-y-7">
                         <li>
